+++ packages/framework/lib/repository/data_model.dart
@@ -1,0 +1,5 @@
+abstract class DataModel {
+  int? get id;
+
+  Map<String, dynamic> toJson();
+}