+++ packages/framework/test/mocks.mocks.dart
@@ -2,8 +2,8 @@
 // in flrx/test/mocks.dart.
 // Do not manually edit this file.
 
-import 'package:flrx/components/logger/base_logger.dart' as _i3;
-import 'package:mockito/mockito.dart' as _i1;
+import 'package:flrx/components/logger/base_logger.dart' as i3;
+import 'package:mockito/mockito.dart' as i1;
 
 // ignore_for_file: comment_references
 // ignore_for_file: unnecessary_parenthesis
@@ -11,4 +11,4 @@ import 'package:mockito/mockito.dart' as _i1;
 /// A class which mocks [Logger].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockLogger extends _i1.Mock implements _i3.Logger {}
+class MockLogger extends i1.Mock implements i3.Logger {}