+++ packages/flrx_test/CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.0.3
+Fixed: When the store printed an action, it would fail if the action was not a string
+
 ## 0.0.2
 Fixed: TestErrorManager init signature